refactor(osb): introduce a ServiceCatalogRepository
The repositories simplify the git interactions and remove file handling
logic from the spring-OSB service implementations.
