+++ src/main/kotlin/io/meshcloud/dockerosb/service/GenericCatalogService.kt
@@ -1,6 +1,5 @@
 package io.meshcloud.dockerosb.service
 
-import io.meshcloud.dockerosb.persistence.GitOperationContext
 import io.meshcloud.dockerosb.persistence.GitOperationContextFactory
 import mu.KotlinLogging
 import org.springframework.cloud.servicebroker.model.catalog.Catalog
@@ -8,8 +7,6 @@ import org.springframework.cloud.servicebroker.model.catalog.ServiceDefinition
 import org.springframework.cloud.servicebroker.service.CatalogService
 import org.springframework.stereotype.Service
 import reactor.core.publisher.Mono
-
-private val log = KotlinLogging.logger { }
 
 @Service
 class GenericCatalogService(
@@ -20,32 +17,25 @@ class GenericCatalogService(
 
   init {
     contextFactory.acquireContext().use { context ->
-      this.cachedCatalog = fetchAndCacheCatalog(context)
+      val repository = context.buildCatalogRepository()
+
+      this.cachedCatalog = repository.getCatalog()
     }
   }
 
-  private class YamlCatalog(
-      val services: List<ServiceDefinition>
-  )
-
-  /**
-   * Fetches the latest service catalog from git
-   */
-  private fun fetchAndCacheCatalog(context: GitOperationContext): Catalog {
-    context.gitHandler.pullFastForwardOnly()
-
-    val catalog = parseCatalog(context)
-    this.cachedCatalog = catalog
-
-    return catalog
-  }
 
   /**
    * When ever the endpoint /v2/catalog is accessed  it pulls the catalog from the git repo
    */
   override fun getCatalog(): Mono<Catalog> {
     contextFactory.acquireContext().use { context ->
-      val catalog = fetchAndCacheCatalog(context)
+
+      context.attemptToRefreshRemoteChanges()
+
+      val repository = context.buildCatalogRepository()
+      val catalog = repository.getCatalog()
+
+      this.cachedCatalog = catalog
 
       return Mono.just(catalog)
     }
@@ -64,22 +54,4 @@ class GenericCatalogService(
 
     return Mono.just(serviceDefinition)
   }
-
-  companion object {
-    fun parseCatalog(context: GitOperationContext): Catalog {
-      val statusYml = context.gitHandler.fileInRepo("catalog.yml")
-
-      if (!statusYml.isFile) {
-        log.error { "Could not read catalog.yml file from '${statusYml.absolutePath}'. Will start with an empty catalog." }
-        return Catalog.builder().build()
-      }
-
-      val catalog = context.yamlHandler.readObject(statusYml, YamlCatalog::class.java)
-
-      return Catalog.builder()
-          .serviceDefinitions(catalog.services)
-          .build()
-    }
-  }
-
 }
